refactor: move queries and types to separate files
